requirements.txt
acrcloud==2.0
annotated-types==0.7.0
anyio
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
sniffio==1.3.1
starlette
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
fastapi
uvicorn==0.24.0
shazamio
aiofiles==23.2.1
